Extend expr_evaluate() to handle floats & Modify expr_print() to display minimum number of parens

+++ compiler/src/AST/expr.c
@@ -13,7 +13,7 @@ struct expr * expr_create ( expr_t kind,
 	return e;
 }
 
-struct expr * expr_create_value ( int value )
+struct expr * expr_create_value ( float value )
 {
 	struct expr *e = expr_create(EXPR_VALUE, 0, 0);
 	e->value = value;
